backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
python-docx>=1.1.0
PyPDF2>=3.0.1
openpyxl>=3.1.2
pandas>=2.2.0
scikit-learn>=1.6.0
numpy>=2.1.0
httpx>=0.26.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
